Added Comments and organized files
